moved issue folder to overall bathroom folder
