+++ src/main/java/com/nighthawk/spring_portfolio/mvc/bathroom/Issue.java
@@ -1,8 +1,9 @@
-package com.nighthawk.spring_portfolio.mvc.issue;
+package com.nighthawk.spring_portfolio.mvc.bathroom;
+import java.util.ArrayList;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
-import java.util.ArrayList;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import lombok.AllArgsConstructor;
