+++ src/main/java/com/nighthawk/spring_portfolio/mvc/bathroom/IssueApiController.java
@@ -1,4 +1,4 @@
-package com.nighthawk.spring_portfolio.mvc.issue;
+package com.nighthawk.spring_portfolio.mvc.bathroom;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
