+++ src/main/java/com/nighthawk/spring_portfolio/mvc/bathroom/IssueJPARepository.java
@@ -1,6 +1,5 @@
-package com.nighthawk.spring_portfolio.mvc.issue;
+package com.nighthawk.spring_portfolio.mvc.bathroom;
 
-import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
